backend/requirements.txt
Flask==2.3.2
SQLAlchemy==2.0.36
alembic==1.11.1
psycopg2-binary==2.9.6
python-dotenv==1.0.1
pytest==8.3.3